adjusted some files and trained for the first time, see output folder

+++ src/solver.py
@@ -58,7 +58,7 @@ class Solver(object):
             # TRAINING
 
             for i, (inputs, targets) in enumerate(train_loader, 1):
-                inputs, targets = Variable(inputs), Variable(targets)
+                inputs, targets = Variable(inputs.float()), Variable(targets.long())
                 if model.is_cuda:
                     inputs, targets = inputs.cuda(), targets.cuda()
 
@@ -93,7 +93,7 @@ class Solver(object):
             val_scores = []
             model.eval()
             for inputs, targets in val_loader:
-                inputs, targets = Variable(inputs), Variable(targets)
+                inputs, targets = Variable(inputs.float()), Variable(targets.long())
                 if model.is_cuda:
                     inputs, targets = inputs.cuda(), targets.cuda()
